requirements.txt
langchain-openai
langchain-ollama
langchain-huggingface
langchain-chroma
langchain-core
langchain
langchain-community
langgraph
openpyxl
python-dotenv
python-json-logger
pyodbc
streamlit
fastapi
uvicorn
pytest
fastapi[standard]
sentence-transformers
sqlglot